requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20
chromadb>=0.4.22
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
